Updated TypeAheadWithOther component to pass in suggestions, and updated pages that use it
