+++ app/[locale]/styleguide/page.tsx
@@ -62,7 +62,7 @@ import {
 import { BrandColor, Example, handleDelete } from "./sg-components";
 
 import TypeAheadInput from '@/components/TypeAheadInput';
-import TypeAheadWithOther from '@/components/Form/TypeAheadWithOther';
+import { TypeAheadWithOther, useAffiliationSearch } from '@/components/Form/TypeAheadWithOther';
 import {
   useAffiliationsLazyQuery,
   AffiliationsDocument,
@@ -89,37 +89,10 @@ import {
 
 function Page() {
   const [otherField, setOtherField] = useState(false);
-  const [suggestions, setSuggestions] = useState<SuggestionInterface[]>([]);
-  const [fetchAffiliations] = useAffiliationsLazyQuery();
-
+  const { suggestions, handleSearch } = useAffiliationSearch();
   const toastState = useToast(); // Access the toast state from context
   const router = useRouter();
 
-  const handleSearch = useCallback(
-    debounce(async (term: string) => {
-      if (!term) {
-        setSuggestions([]);
-        return;
-      }
-
-      const { data } = await fetchAffiliations({
-        variables: { name: term.toLowerCase() },
-      });
-
-      if (data?.affiliations?.items) {
-        const affiliations = data.affiliations.items
-          .filter((item): item is NonNullable<typeof item> => item !== null)
-          .map((item) => ({
-            id: String(item.id),
-            displayName: item.displayName,
-            uri: item.uri,
-          }));
-        console.log("AFFILIATIONS", affiliations);
-        setSuggestions(affiliations);
-      }
-    }, 300),
-    [fetchAffiliations]
-  );
   // NOTE: This text is just for testing the richtext editors
   const html = String.raw;
   const richtextDefault = html`
